devcontainer: Use pico_setup.sh to install openocd
To use openocd together with RPi pico_w, one cannot use the main branch
of the repository. Also, few extra flags are needed to properly
configure openocd.

As a first step, the Dockerfile download the pico_setup.sh script from
the raspberry-pi github and uses that to easily configure the
envrionment.